CommitUser

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,27 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context,UserManager<ApplicationUser> usermanager)
         {
             _context = context;
+            _userManager = usermanager;
+
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items.Where(e => e.User.Id == user.Id).ToList();
 
             return View("Index", model);
         }
@@ -23,12 +29,14 @@ namespace WishList.Controllers
         [HttpGet]
         public IActionResult Create()
         {
+           
             return View("Create");
         }
 
         [HttpPost]
         public IActionResult Create(Models.Item item)
         {
+            item.User = _userManager.GetUserAsync(HttpContext.User).Result;
             _context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
@@ -37,6 +45,11 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            if(user.Id!=item.User.Id)
+            {
+                return Unauthorized();
+            }
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
